src/agents/insurance/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
langchain
langchain-openai
faiss-cpu
numpy
rapidfuzz
python-dotenv
SQLAlchemy
